Remove processing of interface mappings
Processing of interface mappings, and the associated
'physical_interface_mappings' config entry, was inherited from the Linux
Bridge agent code.  In non-Calico OpenStack setups, interface mappings are
used to:

- give a name (e.g. "physnet1") to each subnet that one or more compute
  node is externally connected to

- specify, on each compute node, which of that node's network interfaces
  is connected to that subnet
  (e.g. "physical_interface_mappings = physnet1:eth0")

- specify how that subnet can be partitioned to transport OpenStack data,
  e.g. by using VLANs on it
  (e.g. "network_vlan_ranges = physnet1:100:299").

Calico doesn't need any of this:  (1) It doesn't do any L2 partitioning of
the external network, so doesn't need anything like network_vlan_ranges.
(2) It relies on normal IP routing, set up dynamically by BGP, to decide
which of the external subnets to use for transport OpenStack data.

In addition, and even in the Calico agent, "physical_interface_mappings" is
used - if configured - to generate a unique ID for the agent - one that
distinguishes it from other agents running on other compute nodes.  However
we haven't been configuring physical_interface_mappings in our setup, and
there is a fallback algorithm that generates the ID by enumerating all
Ethernet devices and taking the first one's MAC address.  We can continue
to rely on that fallback, as we already have been.

Therefore we can remove all of the interface_mappings processing.