Remove "Edit on GitHub" link from search results page
The ReadTheDocs theme puts an "Edit on GitHub" link in the top right-hand
corner of every page. This is great, except for the search page, which is
created by a template rather than a *.rst file, so the link to "Edit on GitHub"
is broken on search pages. This commit means that link isn't shown on the
search results page.