Get node IP when creating a pod.
As a precursor to fixing the REDIRECT issue, need to know the IP
for the node that we're creating the pod on.

We'll use this IP in a following commit, to assign to the caliXXX
veth. That way the REDIRECT iptables rule will work for traffic
coming from a caliXXX interface.