Allow containers to reach the host.
Starting in Felix v0.25, the default behaviour is to prevent
endpoints from reaching thier local host IP.  However, this is
required for common containerized environments like Mesos,
where the containerized executor exchanges messages with a
Mesos Slave process on the host.

This change reverts Felix to the pre-0.25 behaviour by setting
a configuration field in etcd.

It also includes an ST for regression testing.