InterfaceWatcher loop: restructure and use IFLA_OPERSTATE
According to
https://www.kernel.org/doc/Documentation/networking/operstates.txt,
IFLA_OPERSTATE is the most convenient overall representation of
whether an interface is ready for use.  If IFLA_OPERSTATE is
IF_OPER_UP, it is; otherwise, not.  Therefore this change only signals
an interface_update message from the InterfaceWatcher, for a newly
notified interface, if that interface's IFLA_OPERSTATE is IF_OPER_UP.

It also refactors the code a bit, to avoid deep netlink message
parsing when we're not going to signal anything anyway, and adds more
debug logging.