Use the kernel's RPF check instead of iptables rules.
Our iptables rules were incorrect, so we required the kernel's
RPF check to be configured or packets could be spoofed.

This change

* explicitly enables the kernel's RPF check on each
  Calico-controlled interface
* explicitly checks that the global RPF value is compatible with
  a per-interface check
* removes our iptables rules that duplicate the RPF check
* adds unit tests for the per-endpoint rule generation code.

It also reverses the sense of the MARK value on a packet so that a
packet has MARK==1 iff a profile chain matches it.  This makes
the logic much simpler to understand and removes the original bug,
which was caused by conflating "packet was not marked as not
matched" and "packet was matched".

Fixes #790.