Remove interface record if it goes down
This ensures that the InterfaceWatcher loop will emit a new
interface_update message if an interface flaps down and then up again.
It may be that the flap causes routes or other interface state to be
lost, so it makes sense for Felix to check and reprogram in that case.