Change env var POLICY_ONLY_CALICO -> CALICO_NETWORKING
CALICO_NETWORKING defaults to true. When set to false, bird and confd are disabled.

The reason for the name change is to avoid confusion in the plugins. In the Kubernetes plugin we use an environment variable called CALICO_POLICY that indicates when the user wants to specify policy for pods. There should be a clear distinction between this environment variable and the calicoctl environment variable that disables bird.