Use standard logging in test code, instead of print
For some reason, test code 'print'ing adds 10 or 20 seconds to each
test suite run, even though we never actually see the printed
information (because tox or one of the other test runner layers
captures it).

So, let's not do that.  But, the existing test code has a lot of
logging-like statements that could well be useful in understanding and
debugging some future test situation.  So let's use proper logging for
those.

(At a guess, eventlet or testr is monkey-patching sys.stdout and
writes are blocking for ~300ms on some print calls.  Possibly some
interaction with our mocking out of time and eventlets monkey
patching.

Found by logging timstamps and doing a binary chop!)

Change-Id: I651588bdcf4a35980a7b82933cf9de6a1abd10b4