Don't automatically install and use Calico DHCP agent
Instead we continue using the Neutron DHCP agent by default.  We will
document that the new Calico DHCP agent is available for people to try
if using Liberty, how to do that, and what its advantages are.  We
will also work on making the Calico DHCP agent work with the previous
Neutron releases that we support, as well as with Liberty and future
Neutron releases.  Once we're confident about all that, we'll switch
to using the Calico DHCP agent by default.

Even though we don't use it by default, the availability of the Calico
DHCP agent for Liberty is a notable new feature, so we flag that with
the Sem-Ver note below.  (This should ideally have been on the commit
that added the Calico DHCP code - but was forgotten at that point.)

Change-Id: I953d6244c9be9e94e225ae3e29438abccc596d4d
Sem-Ver: feature