Support master Neutron code (Mitaka) as well as previous releases
This change adapts networking-calico's code so that it works with the
Neutron master code, as well as continuing to work with Liberty and
previous OpenStack releases.  As we're now very close to the Mitaka
release, working with master Neutron code should also mean working
with the Mitaka release.

Handle removal of deprecated use_namespaces option:

- Don't override use_namespaces, as it no longer exists to be
  overridden.
- Handle new Mitaka NetModel calls that don't allow a use_namespaces
  argument, and set _ns_name attribute to get equivalent effect.
- Use subclassing to override pieces of the DHCP agent implementation
  that assume they own their own namespace (and so delete unknown
  devices, and set the default route).

Adapt for other Neutron changes since Liberty:

- neutron.agent.linux.dhcp's commonutils import changed to common_utils.
- DHCP agent code now requires NetModel.mtu to be a number (or else for
  'advertise_mtu' to be configured False).
- New mtu parameter in RoutedInterfaceDriver.plug_new.
- Adjust test code for different eventlet thread timing.

Change-Id: Ib580639dc341df8a931a947fb3d82f97d656a32b
Sem-Ver: feature