Fix that next-tier action also accepted packets.
Previously, we were setting the accept mark at the start of
the profile chain and then unsetting it at the end of the chain
if no rules matched.  That was correct while the only returning
rule type was "allow", but it meant that when we added
"next-tier" it also implicitly accepted the packet.

Fixes #1014