Fix that a config section called [DEFAULT] was ignored.
In addition, make config file optional.

The config parser treats a section called [DEFAULT] as a special-case
section that is used as default values for other sections. If there
happen to be no other sections then we would miss the config values in
the default section.

Bypass that behaviour and add logging to early startup, enabled via
environment variable.

Fixes #1036.