DHCP agent: don't directly connect different subnets
In Neutron, if a network has multiple subnets, the interpretation is
that those subnets' CIDRs are being used, mixed together, on the same L2
segment, and hence that the subnets' CIDRs are directly connected to
each other.  Therefore the DHCP agent code generates options in the DHCP
response that cause the instance to establish additional directly
connected routes through the instance's NIC.

For example, if an instance is connected to a network with two subnets,
172.17.48.0/20 and 172.16.48.0/20, and the instance's NIC gets an IP
address of 172.17.48.56, the instance will establish routing as:

default via 172.17.48.1 dev eth0
172.16.48.0/20 dev eth0
172.17.48.0/20 dev eth0

In traditional Neutron, with L2 semantics, this is correct.  With Calico
semantics, however, it is not correct: even instances on the same CIDR
should not be assumed to be directly connected, and certainly not
instances on different CIDRs.

Further, the Calico DHCP agent exposes this behaviour more readily than
the standard Neutron DHCP agent would, because of how it coalesces
subnets onto a single network model.  That means that, with the Calico
DHCP agent, subnets are _incorrectly_ modelled as being directly
connected to each other even when they are defined on _different_
Neutron networks.

In this change we fix that by subclassing
_generate_opts_per_subnet (starting from current Neutron stable/mitaka)
and removing the code that generates the directly connected cross-subnet
routes.  We also simplify the code by inlining that all Calico subnets
are what Neutron calls 'isolated'.

Change-Id: If71799e31fe5974ecc9036419d51c42480a6f4a9