Reduce default etcd watcher timeout to 10s and make configurable.
This makes us more responsive to connection failures and reduces the
upper bound on how long we'll wait to process the resync_after_current_poll
flag.

At very low churn rates, this increases the number of etcd requests, but
only to a level that we already support in order to handle higher churn.

At high churn rate, this makes no difference because we'll never hit the
timeout.