Do floating IP DNATs from OUTPUT as well as from PREROUTING
This is needed so that an endpoint can be pinged from its own host using
its floating IP.  One reason for wanting that is that that is what
OpenStack Tempest tests do.  More generally it adds utility without any
downside.

The PREROUTING chain is only traversed for packets that are received on
the compute host from elsewhere; so our use of PREROUTING to implement
floating IPs works for such packets, but not for packets that originate
from the compute host.  To cover the latter, we can traverse the
floating IP DNAT rules from the OUTPUT chain as well.

Note that we also (already) do floating IP-related SNATs in the
POSTROUTING chain.  For those to operate correctly, the DNATs have to
happen first.  (And happily they do:
https://upload.wikimedia.org/wikipedia/commons/3/37/Netfilter-packet-flow.svg)