Packaging: remove Python code from 'calico-common' package
Continuing to include that code, in a Debian package, causes dependency
issues that are entirely spurious because we actually now use
PyInstaller to pull in all of our Python dependencies, and then wrap the
PyInstaller build into the 'calico-felix' package.  Hence we totally
don't need the Debian machinery also generating dependencies on
corresponding Python debs.

(Specifically, the remaining Python code now depends on grpcio, and it
appears that there is no python-grpcio package available in Ubuntu
Trusty.)

Therefore, this change trims 'calico-common' so that it only now
provides:

- calico-diags

- our BIRD config templates and sample shell scripts.