add flag to use predefined host ASN
In some setups, it may be necessary to have peer using different ASNs (for
example when using unique ASN per rack). When the Calico node container starts,
the preconfigured ASN is overwritten and the node is no longer able to peer. By
preferring the preconfigured value when no ASN is defined, we are able to
survive container restarts.

This is particularly important when using a DaemonSet to deploy the nodes to all
hosts.