DevStack: Use 'calico-1.4' PPA instead of 'kilo-testing'
The DevStack plugin uses a PPA for two reasons:

- for the Calico dependencies that we don't build from source and that
  are not available in Ubuntu

- when running against OpenStack Kilo, to pick up Calico-modified Nova
  and Neutron packages.

Please note, in connection with the latter point, that from Liberty
onwards Calico works with vanilla Nova and Neutron - so that point
really is only needed when running with Kilo.

Ignoring the Nova and Neutron packages that are needed for Kilo point,
most of the PPAs at https://launchpad.net/~project-calico have identical
versions of the dependency packages that Calico needs.  However the
'calico-1.4' PPA has these packages for both Trusty and Xenial, whereas
'kilo-testing' only has them for Trusty.  As OpenStack CI jobs now need
to run on Xenial, we need to start using 'calico-1.4'.

It follows that we are temporarily losing the ability of the
networking-calico DevStack plugin to run against OpenStack Kilo (as
'calico-1.4' does not provide the Nova and Neutron packages that are
needed for Kilo).  However it is also the case that we currently have no
OpenStack CI for networking-calico against Kilo.  I hope to add that in
soon - after completing the transition of the master OpenStack CI to
Xenial - and will address the Kilo PPA point then (either by arranging
for the Kilo CI to run on Trusty, or by adding Xenial packages to
'kilo-testing').

Change-Id: I045be86e49f550139945f27d4cd5992079315ff7