Remove handling for dirty working trees
i.e. for a tree with uncommitted changes (whether staged or not).  These
cause a problem with Debian and RPM package generation - because the
updating of the debian/changelog (or rpm/felix.spec) means that the tree
is dirty by the time that we start generating the RPM (or Debian)
packages.