Document 1.4.x branch release process
As we now have different release processes on different branches, it's
better to have a code-managed RELEASING.md than to try to describe the
process in a central place.  We already now have RELEASING.md on master;
this commit adds the corresponding file for 1.4.x.

Also add missing 1.4.5 release notes to CHANGES.md.