Fix floating IP interception for Kilo
The hooking of _create_floatingip doesn't work for Kilo because Kilo
didn't have a _create_floatingip method.  But it does have
create_floatingip (without the leading underscore), and so also do
Liberty and Mitaka, with exactly the same signature as the underscore
version in Liberty and Mitaka.  So hook create_floatingip instead.

Change-Id: Iefb1f2377b31960a84b0b64c822eb2bf30e3685a