Revert setup.py >=1.8 constraint for pbr
This reverts one of the changes made by commit
7ff3bd47f687f587e9537d4e21b72999944a74f6.

With 'pbr>=1.8' in setup.py, Launchpad cannot build networking-calico
for Trusty, because the Trusty archive only has pbr 0.7 (and I was not
able to find any authoritative backports).  In fact I don't believe
there is anything in networking-calico itself that really requires
pbr>=1.8, so relaxing this should allow networking-calico to build for
Trusty.

On the other hand, we still keep the 'pbr>=1.6' constraint in
requirements.txt, for consistency with the OpenStack constraints system.
requirements.txt is not directly involved when a Debian source package
is built and uploaded to Launchpad, and binary packages then built from
that.

Change-Id: If11be6d0f9fb7fab6c7ca4ba776d323486fb1383