Refactor of confd and k8s backend
As we are planning to migrate off of confd in the long run I refactored
the code to be a bit more like an actual go repo and switched to Glide
for dependency management.

I have also added a very naive k8s backend that accounts for everything
we currently store in k8s, we will need to add more when we populate the
rest.

- Moved the contents of `src/github.com/kelseyhightower/confd/*` into
  the top level directory
- Migrated to Glide for dependencies
- Added k8s backend to pull needed "keys" from k8s
- Created `watch()` implementation
  - Grabs a watch from the k8s backend where it can
  - sleeps for 10 seconds where we aren't yet storing values
  - sleeps for 1 minute and logs a debug message when we aren't
    expecting that prefix