Enable caching K8s binaries for K8s Vagrant install
Added feature to K8s vagrant install where it is possible to
pre-download the K8s binaries and mount them into the VMs.
To do this a script was created that can be used to download the
binaries (either locally or in VMs) and then install them if
specified.  The mounting of the cache directory is only attemped
if the K8S_CACHED env variable is set true.