Stop failing for broken links in v2.0 docs
Specifically, networking-calico docs were just removed from
docs.openstack.org, so
http://docs.openstack.org/developer/networking-calico/ now no longer
exists, but the v2.0 docs include a reference to that.

(As do the v1.5 and v1.6 docs, for which we already regard broken links
as non-blocking.  For v2.1 the doc was updated so as to remove that
link, so this problem finishes with v2.0.)