Added dynamic log level
These changes give the user the possibility to change calico's log level.
The log level can be changed by launching the container with the `LOG_LEVEL` environment variable containing the desired level.

This commit affects the calico.conf.default template to add a placeholder for the log level. It changes the install-cni.sh script and add the sed command to add the log level.

Finally it also provides a default value for the log-level as an env variable in the Dockerfile.