Fixing calico-ipam using hostname instead of nodename
It looks as though we were still using the deprecated `hostname` from
the CNI payload, and not `nodename` which was causing issues assigning
affine blocks if you need to use a `NODENAME` that is different from the Nodes `os.Hostname()`.

We'll check for the old `hostname` value as well similar to how we do in
the CNI plugin.