Fixing memory leak in KDD watchers
When we create a cluster without our CRDs or if we remove one from a
running cluster it will start retrying to watch that non-existent
resource. During this loop we "resync" and then destroy the old
watchers. This process kicks off the leak which stems from somewhere in
client-go, which could be from fragmenting memory by quickly creating
and destroying the watches and underlying channels.

We now only close out watchers that have needed resync, this prevents us
from retrying watches on things that don't need to be stopped.