Adding tests for install-cni.sh
I've added several tests to ensure at least basic functionality of the
install-cni.sh script, using the container package from the Felix repo.

I've also resolved a couple of minor bugs in the script for it not
failing when it should.

Similar to how we do it in the kube-controllers tests, we'll pre-complie
the tests to run them outside a container, and allow it to interact with
Docker.