Fix networking-calico CI (interface.OPTS move)
The test code that was registering interface.OPTS didn't actually need
those options anyway (which are for OVS), so simplest just to delete
that registration line.

Also avoid using sudo when the DevStack plugin runs auto-bird-conf.sh;
instead auto-bird-conf.sh itself uses sudo for just the places that need
it.

Change-Id: If2b6199842a9c1673c936344b32031fb09c097ee