Fixing TLS directory error on install
The option `-e` was only checking that the tls dir existed, but not if
it had any contents. When empty (which it is sometimes) the following
`cp` call would fail. I've added a check to see if the dir exists and
has something in it.

There is a similar fix in the sleep loop at the end checking to see if
we need to update the secrets.