Running install_cni_test.cleanup() with exec.Command()
Currently, the way the felix.containers works is when you run a
container it waits to see it come online by issuing some checks (`docker
ps`, and `docker inspect`) which works fine for long running containers.
On short lived containers, such as the `cleanup()` call, the container
might sometimes exit before it has been verified to be running.

I've changed the `cleanup()` call to use an `exec.Command()` to reach
out to docker instead of using the felix.containers package to get
around this.