CentOS/RHEL 7 itself now provides a new enough version of dnsmasq
Therefore our instruction to run 'yum update', in order to upgrade
dnsmasq, is wrong.

However, this assumes using the latest available version of CentOS/RHEL
7 - currently 7.4 - so we should also remove the suggestion of
downgrading to 7.2 in order to work around the Nova bug in Mitaka and
earlier.  In practice, most new installers should be using Newton or
later, and there is still the manual patching option for earlier
OpenStack releases.  (And it is bad practice to suggest using a version
of CentOS/RHEL that is 2 minor versions behind what is now available.)