Disable Kubernetes' termination grace period for calico/node
In Kubernetes' v1.8, the DaemonSet scheduler was changed to respect the
termination grace period on a pod.  This had a couple of knock-on effects:

- The kubelet started giving calico/node 30s to stop on SIGTERM before
  killing it.  calico/node was using the full 30s because install_cni.sh
  doesn't respond to SIGTERM.

- The scheduler started blocking scheduling of a new calico/node pod
  after a deletion until the kubelet on a node reports the completion
  of the above grace period.

Combining the two changes: the old calico/node would get stuck in the
TERMINATING state for 30s even though bird and felix stopped almost
immediately.  Then, kubelet would have to spot that the termination had
actually finished and report that through the API server, which seemed to
add another 30+s of delay before a new calico/node pod was schedulted.

This change disables both behaviors on the calico/node pod by setting
the termination grace period to 0.  0 is a special-case value that allows
a new pod to be started immediately without waiting for the status to
be reported by kubelet.  In fact, the kubelet still gives the old pod
a short grace period in this case, in order to write out logs etc. so
the behaviour is ideal for Calico.

Fixes https://github.com/projectcalico/felix/issues/1592