Fixing race conditions for global config values
There were a couple spots in `startup.go` that were able to invlove race
conditions if two Nodes were trying to set the values at the same time
in KDD, we've run into this before and have a helper function for
protecting those calls from terminating the startup process when in
reality there is no problem.

Removing the checkConflictError function and inlining the calls to this function.