Allow 10 retries between all hosts, not just 1->3/4/5
We have been seeing flakes in this test, and hypothesis is that that's
because we don't allow enough time for routes to propagate over BGP
after creating workloads.

As the code stood, it was possible for propagation to happen very
quickly from hosts 3/4/5 to 1, but it might not yet have happened on
other peerings; and then the following assert_ip_connectivity, which
used to allow no retries at all, could fail.