'. startup.env' -> '. ./startup.env'
calico/node is using Busybox's sh, and there was a bump in Alpine:3.6
3 days ago, that could have brought in a increment to Busybox.

I couldn't find doc for Busybox sh, but Bash's doc for the 'source'
builtin says:

  source filename [arguments]
         [...] If
         filename does not contain a slash, filenames in PATH are used to  find  the
         directory  containing  filename.  [...]

So it's plausible that Busybox sh has been tightened to match this, and
that our '. startup.env' was always wrong and only worked by accident.