Delete UDP conntrack state before rechecking connectivity
This is an attempt to address calico/node ST flakes like the one here:
https://semaphoreci.com/calico/calico/branches/master/builds/1238

I guess this is caused by leftover conntrack state.  The overall flow in
the test that failed is:

1. Create 6 workloads in 2 disjoint groups, and check
connectivity.  (Workloads can connect to others in same group, but not
in the other group.)

2. Modify profile rules so that there is connectivity between the
groups, and check that.

3. Restore normal profile rules, and check normal (=> disjoint)
connectivity again.

The failure is seen in part (3), specifically because a UDP
connection (echo hello | nc -u -w1 $1 69 | grep hello) succeeds when the
test expects it to fail.

Adding the nc -r option might help, although it sounds like it would not
theoretically eliminate all possibility of source port reuse:

     -r      Specifies that source and/or destination ports should be
      chosen randomly instead of sequentially within a range or in the
      order that the system assigns them.

Alternatively we can flush the conntrack state before step 3, and that's
what this commit does.