Allow for ready window in health test FV
Since we allowed Felix to report itself as live while waiting for the
datastore, we now have a small 'ready' window between the end of the
config loop and starting up the dataplane (in the case, as in this test,
where the dataplane then decides for some reason that it is not ready).

Adjust the test expectation to allow for this.