Document possible effects of etcd compaction
etcd compaction potentially affects calls where we are specifying a
particular etcd revision, but that etcd revision is no longer available
because of etcd compaction.  In networking-calico that means watch(),
put() and delete() calls with a specific mod_revision.  In each case the
window for compaction to make itself felt is relatively small -
basically, between when networking-calico reads some data, that has a
particular revision, and when it then decides to delete or update that,
or to start watching for further changes from that revision.  But the
window still exists.

- For watch(), compaction can cause a watch() call to silently fail; but
  we timeout each watch() call after 10s anyway, so that is OK.  This is
  documented in more detail next to the watch_subtree() call in
  etcdutils.py.

- For put() and delete(), experimentation indicates that compaction does
  not affect the correctness of put/delete calls with a specified
  mod_revision: regardless of compaction, they succeed if there has not
  been another change to the key since that revision, and fail if there
  has.  So there is nothing extra that we need to do in the
  networking-calico code here.

In addition, one might worry about compaction affecting an existing
watch() - i.e. where etcd is compacted after a client has already
started watching a key range from some revision.  But in fact there is
no impact here, because:

- if the keys being watched are changed before the compaction, etcd
  reports events for those keys (as usual) before the compaction

- at the compaction, the key value isn't actually changing, so there is
  no event

- if the keys change after the compaction, etcd reports events for those
  changes as usual.

To put it another way, compaction after a watch has started does not
impact the completeness of the information that etcd is able to tell us
about the watched key range.

Change-Id: I93faf7bdb5fd866629e58d7cc144cfc317fb16de