Whitelist failsafe port response traffic in the raw table only
This fixes a nasty interaction between do-not-track policy and
failsafe ports. Previously, adding do-not-track policy to a host
endpoint resulted in failsafe port traffic being blocked because
one leg of the connection would get conntracked, but the return
leg would hit the do-not-track policy, which would either drop
it (thus breaking the failsafe port) or mark it as NOTRACK,
which results in future outbound packets failing the
--ctstate INVALID test (because conntrack has only seen an
outbound SYN, it isn't expecting the outbound to move past the
SYN stage of the handshake).

This change whitelists the response traffic in the raw table
only so that failsafe ports are now forced to be conntracked,
bypassing any do-not-track policy in both directions. Since
the change is made only in the raw table (and we don't mark
failsafe port traffic with the accepted mark), the response
traffic still has to pass the --ctstate ESTABLISHED test in
the filter table.

Merge pull request #1718 from fasaxc/failsafe-donottrack

(cherry picked from commit da11a06dbab7caa33cee45789b9f5d298e95fa0f)

Whitelist failsafe port response traffic in the raw table only