Ensure that all Calico driver/plugin code logs consistently
It was using a mixture of plain Python 'logging' and OpenStack
'oslo_log', and the practical result of that was missing some logs in
the Neutron server log.  Here we ensure that all code used by the
driver/plugin uses 'oslo_log'.

Also, when running UTs - where we have effectively mocked out 'oslo_log'
- replace all of the relevant modules' LOG attribute by a 'logging'
logger.  That allows us to see logging when working on UTs.

Change-Id: Ic8671eaa46ba9c97bdab6fe557d6005aff035026