Remove docs to disable SELinux and modify QEMU config,
It has been reported that the QEMU config modification is no longer
needed, and our own testing on RHEL/CentOS has now confirmed that, and
it is highly likely that that is equally true on Ubuntu.

It's possible that some SELinux configuration is still needed -
specifically to allow nova-compute and libvirt to access NFS shares
exported from another host - but this falls under general OpenStack
setup and is not specifically related to Calico; so we now delete that
from this doc too.