Configure model FV hosts to drop forwarded traffic by default
Configure our model hosts to drop forwarded traffic by default.  Modern
Kubernetes/Docker hosts now have this setting, and the consequence is
that whenever Calico policy intends to allow a packet, it must
explicitly ACCEPT that packet, not just allow it to pass through
cali-FORWARD and assume it will be accepted by the rest of the chain.
Establishing that setting in this FV allows us to test that.