NodeController simplified/stateless
This commit changes the node controller to close race conditions
where we might delete a Calico node for reasons other than a
delete of the corresponding Kubernetes node.

It simplifies the controller to remove caches and instead
runs a simple sync routine that queries the data sources.

It also removes the FV test that expects Calico Nodes that
are written without a corresponding K8s node to be removed.
This is a change in behavior; we don't actually need to do
this because startup.go will always ensure the K8s node
exists before writing the Calico node.

Signed-off-by: Spike Curtis <spike@tigera.io>