Fix ifacemonitor UT flake
The test code's intention is to test when a deleted link is detected in
a resync.  However there was a race, and the following was also
possible:

1. Kick off a resync.
2. Kick off link deletion.
3. Resync queries links and finds that the link is question is still
present.
4. Link deletion now proceeds, and state for that link deleted.
5. Resync queries addresses for that link and finds that there aren't
any.

resulting in an addr callback with set.New(), instead of an addr
callback with nil.

We can now use delLinkNoSignal to ensure that the original intended
scenario reliably happens, so that's what this change does.