Use JSON instead of YAML as calicoctl input
This is motivated by a calico/node ST flake in which:

- one of the nodes (DockerHosts) gets a name with a leading 0e,
  specifically 0e5524752632

- there's a step in the test that writes YAML including that name, and
  when it does this the name is not quoted:

  apiVersion: projectcalico.org/v3
  kind: Node
  metadata:
    labels:
      routeReflectorClusterID: 224.0.0.2
    name: 0e5524752632
  ...

- calicoctl reads that YAML and parses that value as a number, 0 (i.e. 0
  x 10^5524752632 = 0).

I believe this is a mismatch between the test code using PyYAML, which
implements YAML 1.1, and go-yaml-wrapper implementing YAML 1.2.
0e5524752632 is not allowed as a number form in YAML 1.1, so it's
reasonable for YAML 1.1 to write that (knowing that it is a string)
without any quoting.

Options include:
1. Changing test code to use ruamel.yaml instead of PyYAML.  ruamel.yaml
   speaks YAML 1.2, so I believe it would quote in this case.
2. Changing test code to write JSON for calicoctl instead of YAML.

In this change, I've opted for (2), and have changed all cases in the ST
except where YAML is read from calicoctl and then reapplied without
being parsed and regenerated inbetween.